first

+++ src/pages/Dashboard/MyOrders/MyOrders.js
@@ -1,5 +1,6 @@
 import { Button } from '@mui/material';
 import React, { useEffect, useState } from 'react';
+import './MyOrders.css'
 // import Service from '../Service/Service';
 const MyOrders = () => {
     const email = sessionStorage.getItem('email') 
@@ -11,34 +12,53 @@ const MyOrders = () => {
         .then(data=>setServices(data))
     },[control])
     // // Handle Delete 
-    const handleDelete = (id)=>{
-        console.log(id)
-        fetch(`http://localhost:5000/deleteOrder/${id}`,{
-            method : 'DELETE',
-        })
-        .then(res=>res.json()) 
-        // .then(data=>console.log(data))
-        .then(data=>{
-            if(data.deletedCount){
-                setControl(!control)
-            }
-            console.log(data)
-        })
+    // const handleDelete = (id)=>{
+    //     console.log(id)
+    //     fetch(`http://localhost:5000/deleteOrder/${id}`,{
+    //         method : 'DELETE',
+    //     })
+    //     .then(res=>res.json()) 
+    //     // .then(data=>console.log(data))
+    //     .then(data=>{
+    //         if(data.deletedCount){
+    //             setControl(!control)
+    //         }
+    //         console.log(data)
+    //     })
         
-    }
+    // }
+     const handleDelete = (id)=>{
+        const proceed = window.confirm('Are you sure, you want to delete?') 
+        if(proceed){
+    
+            fetch(`http://localhost:5000/deleteOrder/${id}`,{
+                method : 'DELETE',
+            })
+            .then(res=>res.json()) 
+            // .then(data=>console.log(data))
+            .then(data=>{
+                if(data.deletedCount){
+                    setControl(!control)
+                }
+                console.log(data)
+            })
+            
+        }
+     }
+    
+
     return (
         <div>
-            <h2>My Orders</h2>
+            <h2 className='mb-3'>My Orders {services.length}</h2>
             <div className='services-container'>
-            
 
-               {
+            {
                    services?.map(service=>
                     <div>
                         <div className="container">
                             <div className="row">
                                 <div className="col">
-                                <div className='card w-50 mb-4'>
+                                <div className='card  mb-4'>
         <img style={{width : "50%", margin : "0 auto"}} src={service?.img} className="card-img-top" alt="..."/>
         <div className="card-body">
           <h5 className="card-title">{service?.name}</h5>
